+++ Cargo.toml
@@ -120,7 +120,7 @@ branch = "oxide/omicron"
 #
 # libsodium-sys is a dependency of thrussh, used in `end-to-end-tests`.  We
 # maintain a fork of the sodiumoxide repository to address an illumos build
-# issue.  See the README.oxide.md in the "oxide" branch of our fork for
+# issue.  See the README.oxide.md in the "omicron/oxide" branch of our fork for
 # details.
 #
 [patch.crates-io.libsodium-sys]
